refactor: extract shared BinActionMenuItem component
AddToQueueFromBin and PlayFromBin were 90% identical — same shift-key
logic, same delete-from-bin behavior, same Chip/Typography markup.
Extract the common structure into a parameterized BinActionMenuItem
that accepts icon, label, color, and action props.
